requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1
apscheduler==3.10.4
watchdog==3.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psutil==5.9.6
requests==2.31.0
Pillow==10.1.0